+++ packages/bottom-sheet/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @alfalab/core-components-bottom-sheet
 
+## 8.1.14
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-base-modal@7.2.0
+
+## 8.1.13
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-navigation-bar-private@2.0.14
+
 ## 8.1.12
 
 ### Patch Changes