+++ packages/code-input/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @alfalab/core-components-code-input
 
+## 5.1.3
+
+### Patch Changes
+
+<sup><time>11.09.2026</time></sup>
+
+#### [#2337](https://github.com/core-ds/core-components/pull/2337)
+
+##### CodeInput
+
+- Исправлен скачок вьюпорта при повторном фокусе на ячейке ввода
+
+## 5.1.2
+
+### Patch Changes
+
+<sup><time>08.09.2026</time></sup>
+
+#### [#2344](https://github.com/core-ds/core-components/pull/2344)
+
+##### CodeInput
+
+- Исправлено аварийное закрытие вкладки браузера при размонтировании компонента во время ожидания SMS-кода (воспроизводилось в Яндекс Браузере)
+
 ## 5.1.1
 
 ### Patch Changes