+++ packages/select-with-tags/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @alfalab/core-components-select-with-tags
 
+## 10.1.22
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-bottom-sheet@8.1.14
+- @alfalab/core-components-select@19.2.9
+
+## 10.1.21
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-select@19.2.8
+
+## 10.1.20
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-select@19.2.7
+
+## 10.1.19
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-select@19.2.6
+- @alfalab/core-components-bottom-sheet@8.1.13
+
+## 10.1.18
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-tag@10.3.0
+- @alfalab/core-components-select@19.2.5
+
 ## 10.1.17
 
 ### Patch Changes