+++ packages/tab-bar/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @alfalab/core-components-tab-bar
 
+## 4.0.12
+
+### Patch Changes
+
+#### Обновлены зависимости
+
+- @alfalab/core-components-typography@6.0.11
+
 ## 4.0.11
 
 ### Patch Changes